requirements.txt
python-telegram-bot==20.6
pandas
openpyxl
aiosqlite
Flask